requirements.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backcall==0.1.0
beautifulsoup4==4.11.1
boto3==1.11.9
botocore==1.14.9
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.3
ConfigArgParse==1.5.3
cryptography==37.0.4
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.8.0
geventhttpclient==1.5.3
greenlet==1.1.2
h11==0.13.0
idna==3.2
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.16.0
Jinja2==3.0.2
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
locust==2.4.0
MarkupSafe==2.0.1
matplotlib==3.2.1
msgpack==1.0.2
numpy==1.18.2
outcome==1.2.0
pandas==1.0.3
parso==0.6.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.3
psutil==5.8.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.5.2
pyOpenSSL==22.0.0
pyparsing==2.4.6
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==302
pyzmq==22.3.0
requests==2.26.0
roundrobin==0.0.2
s3transfer==0.3.2
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==4.3.0
six==1.14.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
traitlets==4.3.3
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.10
wcwidth==0.1.8
Werkzeug==2.0.2
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
